Data: regenerating based on 801429a267fc626db16d12f8f653c09c61e7f5e9

+++ data/Pandora.Definitions.ResourceManager/StreamAnalytics/v2020_03_01/StreamingJobs/Model-OutputProperties.cs
@@ -28,7 +28,7 @@ internal class OutputPropertiesModel
     public SerializationModel? Serialization { get; set; }
 
     [JsonPropertyName("sizeWindow")]
-    public float? SizeWindow { get; set; }
+    public int? SizeWindow { get; set; }
 
     [JsonPropertyName("timeWindow")]
     public string? TimeWindow { get; set; }
